# docs/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: CBSC Strategy Management API
  description: |
    CBSC量化交易策略管理系統API文檔

    此API提供了完整的策略管理功能，包括：
    - 策略的創建、讀取、更新、刪除
    - 策略執行和監控
    - 實時市場數據獲取
    - 回測功能
    - 用戶管理

    # 認證
    API使用JWT Bearer Token進行認證。請在Authorization header中包含有效的token。

    # 錯誤處理
    API使用標準HTTP狀態碼。錯誤響應格式如下：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "錯誤描述",
        "details": {}
      },
      "timestamp": "2025-01-19T10:00:00Z"
    }
    ```
  version: 2.0.0
  contact:
    name: CBSC API Support
    email: api-support@cbsc.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cbsc.example.com/v2
    description: Production server
  - url: https://staging-api.cbsc.example.com/v2
    description: Staging server
  - url: http://localhost:3004/v2
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: 用戶登錄
      description: 用戶使用用戶名和密碼登錄，返回JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: 用戶名
                  example: "john.doe"
                password:
                  type: string
                  description: 密碼
                  format: password
                  example: "securePassword123"
                remember_me:
                  type: boolean
                  description: 是否記住登錄狀態
                  default: false
      responses:
        '200':
          description: 登錄成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        description: JWT access token
                      token_type:
                        type: string
                        example: "bearer"
                      expires_in:
                        type: integer
                        description: Token有效期（秒）
                      refresh_token:
                        type: string
                        description: Refresh token
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: 刷新Token
      description: 使用refresh token獲取新的access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: Refresh token
      responses:
        '200':
          description: Token刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      expires_in:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Strategy endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, trader, analyst, viewer]
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        timestamp:
          type: string
          format: date-time
  responses:
...
    Unauthorized:
      description: 未授權或token無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: 請求頻率超過限制
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...